nea_work/requirements.txt
Django==4.2.16
psycopg2-binary==2.9.9
openpyxl==3.1.2
sqlparse==0.5.5
asgiref==3.11.1
python-decouple==3.8
et-xmlfile==2.0.0